requirements/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
